Refactor MainWindow class out into its own module
