+++ RimSort.py
@@ -1,40 +1,32 @@
 import sys
 
-from multiprocessing import current_process, freeze_support, set_start_method
+from multiprocessing import freeze_support, set_start_method
 
 import os
 from pathlib import Path
 import platform
-from requests.exceptions import HTTPError
 from tempfile import gettempdir
 import traceback
-from typing import Any, Optional
 
-from logger_tt import handlers, logger, setup_logging
+from logger_tt import logger, setup_logging
 from logging import getLogger, WARNING
-from PySide6.QtCore import QSize, QTimer
-from PySide6.QtWidgets import QApplication, QMainWindow, QVBoxLayout, QWidget
+from PySide6.QtWidgets import QApplication
 
-from controller.menu_bar_controller import MenuBarController
-from controller.settings_controller import SettingsController
-from model.settings import Settings
 from util.app_info import AppInfo
-from view.menu_bar import MenuBar
+from view.main_window import MainWindow
 
 SYSTEM = platform.system()
 # Watchdog conditionals
 if SYSTEM == "Darwin":
-    from watchdog.observers import Observer
 
     # Comment to see logging for watchdog handler on Darwin
     getLogger("watchdog.observers.fsevents").setLevel(WARNING)
 elif SYSTEM == "Linux":
-    from watchdog.observers import Observer
 
     # Comment to see logging for watchdog handler on Linux
     getLogger("watchdog.observers.inotify_buffer").setLevel(WARNING)
 elif SYSTEM == "Windows":
-    from watchdog.observers.polling import PollingObserver
+    pass
 
     # Comment to see logging for watchdog handler on Windows
     # This is a stub if it's ever even needed...
@@ -43,185 +35,6 @@ elif SYSTEM == "Windows":
 
 from model.dialogue import show_fatal_error
 from util.proxy_style import ProxyStyle
-from util.watchdog import RSFileSystemEventHandler
-from view.game_configuration_panel import GameConfiguration
-from view.main_content_panel import MainContent
-from view.status_panel import Status
-
-
-class MainWindow(QMainWindow):
-    """
-    Subclass QMainWindow to customize the main application window.
-    """
-
-    def __init__(self, DEBUG_MODE=None) -> None:
-        """
-        Initialize the main application window. Construct the layout,
-        add the three main views, and set up relevant signals and slots.
-        """
-        logger.info("Initializing MainWindow")
-        super(MainWindow, self).__init__()
-
-        # Instantiate the settings model and controller
-        self.settings = Settings()
-        self.settings_controller = SettingsController(model=self.settings)
-
-        # Create the main application window
-        self.DEBUG_MODE = DEBUG_MODE
-        self.init = None  # Content initialization should only fire on startup. Otherwise, this is handled by Refresh button
-        self.version_string = "Alpha-v1.0.6.2-hf"
-
-        # Check for SHA and append to version string if found
-        sha_file = str(Path(os.path.join(data_path, "SHA")).resolve())
-        if os.path.exists(sha_file):
-            with open(sha_file, encoding="utf-8") as f:
-                sha = f.read().strip()
-            self.version_string += f" [Edge {sha}]"
-
-        # Watchdog
-        self.watchdog_event_handler = None
-        self.watchdog_observer = None
-
-        # Setup the window
-        self.setWindowTitle(f"RimSort {self.version_string}")
-        self.setMinimumSize(QSize(1024, 768))
-
-        # Create the window layout
-        app_layout = QVBoxLayout()
-        app_layout.setContentsMargins(0, 0, 0, 0)  # Space from main layout to border
-        app_layout.setSpacing(0)  # Space between widgets
-
-        # Create various panels on the application GUI
-        self.game_configuration = GameConfiguration.instance(
-            DEBUG_MODE=DEBUG_MODE,
-            settings_controller=self.settings_controller,
-            RIMSORT_VERSION=self.version_string,
-        )
-        self.main_content_panel = MainContent(
-            settings_controller=self.settings_controller
-        )
-        self.bottom_panel = Status()
-
-        # Connect the game configuration actions signals to Status panel to display fading action text
-        self.game_configuration.configuration_signal.connect(
-            self.bottom_panel.actions_slot
-        )
-        self.game_configuration.settings_panel.actions_signal.connect(
-            self.bottom_panel.actions_slot
-        )
-        # Connect the actions_signal to Status panel to display fading action text
-        self.main_content_panel.actions_panel.actions_signal.connect(
-            self.bottom_panel.actions_slot
-        )
-        self.main_content_panel.status_signal.connect(self.bottom_panel.actions_slot)
-
-        # Arrange all panels vertically on the main window layout
-        app_layout.addLayout(self.game_configuration.panel)
-        app_layout.addWidget(self.main_content_panel.main_layout_frame)
-        app_layout.addWidget(self.bottom_panel.frame)
-
-        # Display all items
-        widget = QWidget()
-        widget.setLayout(app_layout)
-        self.setCentralWidget(widget)
-
-        self.menu_bar = MenuBar(menu_bar=self.menuBar())
-        self.menu_bar_controller = MenuBarController(
-            view=self.menu_bar, settings_controller=self.settings_controller
-        )
-
-        logger.debug("Finished MainWindow initialization")
-
-    def showEvent(self, event) -> None:
-        # Call the original showEvent handler
-        super().showEvent(event)
-        if not self.init:
-            # HIDE/SHOW FOLDER ROWS BASED ON PREFERENCE
-            if self.settings_controller.settings.show_folder_rows:
-                self.game_configuration.hide_show_folder_rows_button.setText(
-                    "Hide paths"
-                )
-            else:
-                self.game_configuration.hide_show_folder_rows_button.setText(
-                    "Show paths"
-                )
-            # set visibility
-            self.game_configuration.game_folder_frame.setVisible(
-                self.settings_controller.settings.show_folder_rows
-            )
-            self.game_configuration.config_folder_frame.setVisible(
-                self.settings_controller.settings.show_folder_rows
-            )
-            self.game_configuration.local_folder_frame.setVisible(
-                self.settings_controller.settings.show_folder_rows
-            )
-            self.game_configuration.workshop_folder_frame.setVisible(
-                self.settings_controller.settings.show_folder_rows
-            )
-
-    def __initialize_content(self) -> None:
-        self.init = True
-
-        # IF CHECK FOR UPDATE ON STARTUP...
-        if self.settings_controller.settings.check_for_update_startup:
-            self.main_content_panel.actions_slot("check_for_update")
-
-        # REFRESH CONFIGURED METADATA
-        self.main_content_panel._do_refresh(is_initial=True)
-
-        # CHECK USER PREFERENCE FOR WATCHDOG
-        if self.settings_controller.settings.watchdog_toggle:
-            # Setup watchdog
-            self.__initialize_watchdog()
-
-    def __initialize_watchdog(self) -> None:
-        logger.info("Initializing watchdog FS Observer")
-        # INITIALIZE WATCHDOG - WE WAIT TO START UNTIL DONE PARSING MOD LIST
-        game_folder_path = self.game_configuration.game_folder_line.text()
-        local_folder_path = self.game_configuration.local_folder_line.text()
-        workshop_folder_path = self.game_configuration.workshop_folder_line.text()
-        self.watchdog_event_handler = RSFileSystemEventHandler()
-        if SYSTEM == "Windows":
-            self.watchdog_observer = PollingObserver()
-        else:
-            self.watchdog_observer = Observer()
-        if game_folder_path and game_folder_path != "":
-            self.watchdog_observer.schedule(
-                self.watchdog_event_handler,
-                game_folder_path,
-                # recursive=True,
-            )
-        if local_folder_path and local_folder_path != "":
-            self.watchdog_observer.schedule(
-                self.watchdog_event_handler,
-                local_folder_path,
-                # recursive=True,
-            )
-        if workshop_folder_path and workshop_folder_path != "":
-            self.watchdog_observer.schedule(
-                self.watchdog_event_handler,
-                workshop_folder_path,
-                # recursive=True,
-            )
-        # Connect watchdog to our refresh button animation
-        self.watchdog_event_handler.file_changes_signal.connect(
-            self.main_content_panel._do_refresh_animation
-        )
-        # Connect main content signal so it can stop watchdog
-        self.main_content_panel.stop_watchdog_signal.connect(self.__shutdown_watchdog)
-        # Start watchdog
-        try:
-            self.watchdog_observer.start()
-        except Exception as e:
-            logger.warning(
-                f"Unable to initialize watchdog Observer due to exception: {e.__class__.__name__}"
-            )
-
-    def __shutdown_watchdog(self) -> None:
-        if self.watchdog_observer and self.watchdog_observer.is_alive():
-            self.watchdog_observer.stop()
-            self.watchdog_observer.join()
-            self.watchdog_observer = None
 
 
 def handle_exception(exc_type, exc_value, exc_traceback):
